% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermDocumentMatrix_classes.R
\name{TermDocumentMatrix_classes}
\alias{TermDocumentMatrix_classes}
\title{TermDocumentMatrix_classes}
\usage{
TermDocumentMatrix_classes(symbol)
}
\description{
TermDocumentMatrix_classes description
}
